Update patch set 1
Patch Set 1:

I compiled it on my phone with this comment and this does fix the current invalid preferred network options. The current preferred network options in hlte is completely a mess with invalid options and force close on selection. preferred_network_mode_choices_cm needs to be added because of an modified array in another recent comment.

Patch-set: 1