Update patch set 1
Patch Set 1:

For me this does enable basic GPS, where it will see satellites, however, I only see 3-4 and the accuracy is terrible.

Need to discuss with the team how best to customize this file as well as continue debugging to see why we're not seeing a normal amount of satellites (16-20+).

Patch-set: 1