Update patch set 1
Patch Set 1: -Code-Review

i could never figure out gerrit so i could never upload that patch..i had given links to it in another commit a while back but no one did anything with it...i tried to reach noob on xda and never got a reply from him either. just bothers me a little bit knowing this is mine cause i had never updated the baseband from the first i did it and it still has the same one in this patchset which is the one the phone shipped with...thats just how i know this was my work.

Patch-set: 1