Update patch set 1
Patch Set 1:

right. I see what you mean. i intended to do that,  does the phone work without the right sim inserted? verizon and sprint have  no choice but to go back to  to their home carrier which mean removing the battery and removing the sim. if it is still in cdma mode, it will still show the option unless you manually change the network selection option because the numeric is still in the build props, if you still have it

Patch-set: 1