Update patch set 2
Patch Set 2:

Yes, the samsung device cleanup that was recently done seems to have changed that.

I don't have a hlte so I can't test for you, but it looks like the cleanup added

av.offload.enable=true

to msm8974-common.  This seems to be fine on mondrianwifi which shares msm8974-common.

Therfore, either we need to set it to false in hlte or move it up to the device and set it to true in mondrianwifi.

Patch-set: 2
Reviewer: Christopher R. Palmer <6351@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0