Update patch set 2
Patch Set 2:

> Abandoned
> 
> HCE works perfectly fine on my hltepsr, which shares the nfc firmware, chip, and configs as hlte (it's unified). I made 3 separate purchases at 3 different stores, and it worked perfectly fine at all of them, no issues whatsoever.

Well, good for you, but for a lot of other people it's broken without the patch. Confirmed on many occasions on xda-developers forum. Vidwhal, maintainer of Ressurection Remix, also confirms that this fixes HCE for him. Fix comes from bacon and klte (for klte it was even added by haggertk), which share the same nfc chip and were not working without it.

Here is the link to original change in LA code: https://review.lineageos.org/c/161917

Patch-set: 2